split loaders into separate files + add logging
